refactor: 강좌 관련 엔티티 course 폴더로 이동
