+++ src/main/java/com/gongkademy/domain/board/Comment.java
@@ -1,5 +1,8 @@
-package com.gongkademy.domain;
+package com.gongkademy.domain.board;
 
+
+import com.gongkademy.domain.BaseTimeEntity;
+import com.gongkademy.domain.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -16,24 +19,27 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Register extends BaseTimeEntity {
-
+public class Comment extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="register_id")
+    @Column(name="comment_id")
     private Long id;
+
+    @Column(nullable = false, length = 1_000)
+    private String content;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(nullable = false, name="member_id")
     private Member member;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(nullable = false, name="course_id")
-    private Course course;
+    @JoinColumn(nullable = false, name="board_id")
+    private Board board;
 
     @Builder
-    private Register(Member member, Course course) {
+    private Comment(String content, Member member, Board board) {
+        this.content = content;
         this.member = member;
-        this.course = course;
+        this.board = board;
     }
 }
