+++ src/main/java/com/gongkademy/domain/board/Question.java
@@ -1,6 +1,6 @@
 package com.gongkademy.domain.board;
 
-import com.gongkademy.domain.Lecture;
+import com.gongkademy.domain.course.Lecture;
 import com.gongkademy.domain.Member;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
@@ -8,11 +8,9 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
 @DiscriminatorValue("Question")
