+++ src/main/java/com/gongkademy/domain/course/Play.java
@@ -1,7 +1,7 @@
-package com.gongkademy.domain;
+package com.gongkademy.domain.course;
 
-
-import com.gongkademy.domain.board.Board;
+import com.gongkademy.domain.BaseTimeEntity;
+import com.gongkademy.domain.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -18,27 +18,32 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Comment extends BaseTimeEntity {
+public class Play extends BaseTimeEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="comment_id")
+    @Column(name="play_id")
     private Long id;
 
-    @Column(nullable = false, length = 1_000)
-    private String content;
+    @Column(nullable = false)
+    private int lastPlayedTime;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(nullable = false, name="member_id")
     private Member member;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(nullable = false, name="board_id")
-    private Board board;
+    @JoinColumn(nullable = false, name="lecture_id")
+    private Lecture lecture;
 
     @Builder
-    private Comment(String content, Member member, Board board) {
-        this.content = content;
+    private Play(int lastPlayedTime, Member member, Lecture lecture) {
+        this.lastPlayedTime = lastPlayedTime;
         this.member = member;
-        this.board = board;
+        this.lecture = lecture;
+    }
+
+    public void changeLastPlayedTime(int lastPlayedTime){
+        this.lastPlayedTime = lastPlayedTime;
     }
 }
