+++ src/main/java/com/gongkademy/repository/CommentRepository.java
@@ -1,6 +1,6 @@
 package com.gongkademy.repository;
 
-import com.gongkademy.domain.Comment;
+import com.gongkademy.domain.board.Comment;
 import java.util.List;
 import java.util.Optional;
