+++ src/main/java/com/gongkademy/repository/CourseRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.gongkademy.repository;
 
-import com.gongkademy.domain.Course;
+import com.gongkademy.domain.course.Course;
 import jakarta.persistence.EntityManager;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
