+++ src/main/java/com/gongkademy/repository/LectureRepository.java
@@ -1,6 +1,6 @@
 package com.gongkademy.repository;
 
-import com.gongkademy.domain.Lecture;
+import com.gongkademy.domain.course.Lecture;
 import java.util.List;
 import java.util.Optional;
