+++ src/main/java/com/gongkademy/repository/LectureRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.gongkademy.repository;
 
-import com.gongkademy.domain.Lecture;
+import com.gongkademy.domain.course.Lecture;
 import jakarta.persistence.EntityManager;
 import java.util.List;
 import java.util.Optional;
