+++ src/test/java/com/gongkademy/repository/CourseRepositoryTest.java
@@ -2,7 +2,7 @@ package com.gongkademy.repository;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import com.gongkademy.domain.Course;
+import com.gongkademy.domain.course.Course;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
