+++ src/test/java/com/gongkademy/service/CourseServiceTest.java
@@ -4,13 +4,12 @@ import static com.gongkademy.exception.ErrorCode.REGISTERED_COURSE;
 import static com.gongkademy.exception.ErrorCode.REGISTER_NOT_FOUND;
 import static org.junit.jupiter.api.Assertions.*;
 
-import com.gongkademy.domain.Course;
-import com.gongkademy.domain.Lecture;
+import com.gongkademy.domain.course.Course;
+import com.gongkademy.domain.course.Lecture;
 import com.gongkademy.domain.Member;
-import com.gongkademy.domain.Play;
-import com.gongkademy.domain.Register;
+import com.gongkademy.domain.course.Play;
+import com.gongkademy.domain.course.Register;
 import com.gongkademy.exception.CustomException;
-import com.gongkademy.exception.ErrorCode;
 import com.gongkademy.repository.RegisterRepository;
 import com.gongkademy.service.dto.CourseDetailResponse;
 import com.gongkademy.service.dto.LectureDetailResponse;
